backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL
# pymysql==1.1.0  # MySQL (可选)

# Redis
redis==5.0.1
hiredis==2.3.2

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# 安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP 客户端
httpx==0.26.0

# AI Provider
openai==1.10.0

# 工具
python-dateutil==2.8.2
pytz==2023.3
Pillow==10.2.0  # 图片处理（用于测试）

# 开发工具
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0
